+++ flaredb/src/engine/mod.rs
@@ -1,4 +1,3 @@
 pub mod coders;
 pub mod executor;
 pub mod harness;
-pub mod store;
