+++ flaredb/src/store/mod.rs
@@ -1,0 +1,9 @@
+pub mod element_store;
+pub mod record;
+
+pub(crate) const ELEMENT_ID_COLUMN: &str = "element_id";
+pub(crate) const COLLECTION_COLUMN: &str = "collection";
+pub(crate) const KEY_COLUMN: &str = "key";
+pub(crate) const VALUE_COLUMN: &str = "value";
+pub(crate) const RECORD_TYPE_METADATA_KEY: &str = "flare.record_type";
+pub(crate) const TABLE_NAME: &str = "table_name";
